Add files via upload
A short Python project which allows customers to make a request of the movie they would want to watch. If the movie exists, a seat is reserved for them and the number of available seats is automatically reduced by 1. However, the client is informed appropriately if the movie does not exist. The purpose is to apply new knowledge to Python lists and functions.