Add files via upload
In this project, I applied my knowledge of while loop and If...else to create a check-in system for my school. If your name is found in the system, it welcomes you. However, if your name isn't, it asks if you are new and would want to join. It doubles as a security system.